+++ example/lib/gallery/gallery_page.dart
@@ -7,7 +7,7 @@ class GalleryPage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return NesScaffold(
+    return NesExampleScaffold(
       child: SingleChildScrollView(
         child: Padding(
           padding: const EdgeInsets.symmetric(horizontal: 32),