+++ lib/src/widgets/widgets.dart
@@ -21,6 +21,7 @@ export 'nes_pressabled.dart';
 export 'nes_pulser.dart';
 export 'nes_running_text.dart';
 export 'nes_running_text_lines.dart';
+export 'nes_scaffold.dart';
 export 'nes_scrollbar.dart';
 export 'nes_section_header.dart';
 export 'nes_selection_list.dart';