+++ widgetbook/lib/widgetbook/widgetbook.directories.g.dart
@@ -12,8 +12,9 @@
 import 'package:widgetbook/widgetbook.dart' as _i1;
 import 'package:widgetbook_app/widgetbook/use_cases/buttons.dart' as _i2;
 import 'package:widgetbook_app/widgetbook/use_cases/checkboxes.dart' as _i3;
-import 'package:widgetbook_app/widgetbook/use_cases/containers.dart' as _i5;
-import 'package:widgetbook_app/widgetbook/use_cases/running_texts.dart' as _i4;
+import 'package:widgetbook_app/widgetbook/use_cases/containers.dart' as _i6;
+import 'package:widgetbook_app/widgetbook/use_cases/dropdown_menus.dart' as _i4;
+import 'package:widgetbook_app/widgetbook/use_cases/running_texts.dart' as _i5;
 
 final directories = <_i1.WidgetbookNode>[
   _i1.WidgetbookFolder(
@@ -52,17 +53,24 @@ final directories = <_i1.WidgetbookNode>[
         ),
       ),
       _i1.WidgetbookLeafComponent(
-        name: 'NesRunningText',
+        name: 'NesDropdownMenu',
         useCase: _i1.WidgetbookUseCase(
           name: 'default',
           builder: _i4.normal,
         ),
       ),
       _i1.WidgetbookLeafComponent(
+        name: 'NesRunningText',
+        useCase: _i1.WidgetbookUseCase(
+          name: 'default',
+          builder: _i5.normal,
+        ),
+      ),
+      _i1.WidgetbookLeafComponent(
         name: 'NesRunningTextLines',
         useCase: _i1.WidgetbookUseCase(
           name: 'default',
-          builder: _i4.lines,
+          builder: _i5.lines,
         ),
       ),
       _i1.WidgetbookFolder(
@@ -73,19 +81,19 @@ final directories = <_i1.WidgetbookNode>[
             useCases: [
               _i1.WidgetbookUseCase(
                 name: 'default',
-                builder: _i5.normal,
+                builder: _i6.normal,
               ),
               _i1.WidgetbookUseCase(
                 name: 'with corner inner square painter',
-                builder: _i5.cornerInnerSquare,
+                builder: _i6.cornerInnerSquare,
               ),
               _i1.WidgetbookUseCase(
                 name: 'with label',
-                builder: _i5.label,
+                builder: _i6.label,
               ),
               _i1.WidgetbookUseCase(
                 name: 'with square corner painter',
-                builder: _i5.squareConer,
+                builder: _i6.squareConer,
               ),
             ],
           )